Pulled in master branch updates

* 'master' of github.com:carothersc/ROSS:
  removed worst line in ROSS history. you're welcome
  Fixed up network-mpi.c with casts.
  Fixed warnings from -Wconversion (and -Wno-sign-conversion).
  Added ross-extern.h to CMakeLists.txt.
  Updates to torus model: added comments for some of the traffic patterns
  updates to torus model (2): making things more configurable, cleaning up some code and adding instructions and common errors in README
  updates to torus model: making things more configurable, cleaning up some code and adding instructions and common errors in README
  AVL behavior change: allow duplicates...
  updates to torus model: making environment compatible with paper runs

+++ ross/hash-quadratic.h
@@ -19,7 +19,7 @@ struct tw_hash
  * hash-quadratic.c
  */
 extern void    *tw_hash_create();
-extern void     tw_hash_insert(void *h, tw_event * event, int pe);
-extern tw_event *tw_hash_remove(void *h, tw_event * event, int pe);
+extern void     tw_hash_insert(void *h, tw_event * event, long pe);
+extern tw_event *tw_hash_remove(void *h, tw_event * event, long pe);
 
 #endif
